requirements.txt
websocket-client
python-telegram-bot
requests
pandas
